Disallow Empty String Values in Mechanism
Prompted by: https://github.com/trailofbits/spf-query/issues/8

When creating a new `Mechanism`, check whether the supplied `value` is
an empty string.  If it is, convert it to `nil` as `#to_s` has logic to
avoid outputting `value` when it's nil.

This works around an issue in [Parslet](https://github.com/kschiess/parslet/issues/155)
where the parser is returning an empty string when it should return
`nil`.